Rename disk temperatures file
