+++ src/middlewared/middlewared/utils/disk_temperatures.py
@@ -1,5 +1,3 @@
-import re
-
 from dataclasses import dataclass
 from typing import Dict, Optional
